requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
pgvector>=0.2.5
sqlalchemy>=2.0.25
pandas>=2.2.0
pypdf>=6.8.0
python-multipart>=0.0.9
python-docx>=1.2.0
openpyxl>=3.1.0
requests>=2.31.0
pydantic>=2.0.0
streamlit>=1.30.0
